Delete 02-de.md
Thank you very much for your contribution. I am very sorry, but unfortunately I have to remove this tutorial as of now, as there are no plans to include German language tutorials as part of the Community tutorials. Again, I am very sorry that this was not mentioned at the time, I suspect this tutorial was missed during the review process back in November. Still, thank you very much for your efforts.